+++ workspace/solutions/rust/src/s0/s9/s1/s2/mod.rs
@@ -1,4 +1,5 @@
 pub mod s_0912_00;
+pub mod s_0912_01;
 pub mod s_0912_test;
 
 pub type TSolution = fn(Vec<i32>) -> Vec<i32>;